V0.5.3
See README.md for changes

+++ srcs/Libft/srcs/Math/ft_abs.c
@@ -6,16 +6,15 @@
 /*   By: siferrar <siferrar@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/11/26 23:29:17 by siferrar     #+#   ##    ##    #+#       */
-/*   Updated: 2019/12/16 22:37:49 by siferrar    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/12/19 17:30:43 by siferrar    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-long ft_abs(int n)
+long	ft_abs(long n)
 {
-
 	n = (n >= 0 ? n : -n);
 	return (n);
 }
